requirements.txt
openai>=1.60.0
anthropic>=0.42.0
pydantic>=2.8.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
requests>=2.32.0
pytest>=9.0.0
